rename version control branch, separate out vcsStatus update imports
