+++ ui-tests/VersionControlComponents.test.tsx
@@ -2,7 +2,7 @@ import { mount, shallow } from "enzyme"
 import { shallowToJson } from "enzyme-to-json"
 import * as React from "react"
 
-import { Branch, BranchNameContainer } from "./../browser/src/UI/components/VersionControl"
+import { Branch, BranchNameContainer } from "./../browser/src/UI/components/VersionControlBranch"
 
 describe("<Branch />", () => {
     const diff = {
